pyproject.toml
[project]
name = "ps3toolbox"
version = "0.1.0"
description = "Tools for PS3 homebrew operations - PS2 ISO encryption and more"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Matheus", email = "" }
]
keywords = ["ps3", "ps2", "encryption", "homebrew", "iso"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Archiving",
]

dependencies = [
    "cryptography>=42.0.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "aiofiles>=25.1.0",
    "thefuzz>=0.22.1",
    "pillow>=12.0.0",
    "aiohttp>=3.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
ps3toolbox = "ps3toolbox.cli:cli"

[project.urls]
Homepage = "https://github.com/yourusername/ps3toolbox"
Repository = "https://github.com/yourusername/ps3toolbox"
Issues = "https://github.com/yourusername/ps3toolbox/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ps3toolbox"]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["ps3toolbox"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src/ps3toolbox --cov-report=term-missing --cov-report=html"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.9",
]
